djangoproject/requirements.txt
amqp==2.5.2
anyjson==0.3.3
apturl==0.5.2
asgiref==3.2.5
asn1crypto==0.24.0
bcrypt==3.1.6
billiard==3.5.0.5
blinker==1.4
Brlapi==0.6.7
celery==4.0.0
certifi==2018.8.24
chardet==3.0.4
CodeConvert==3.0.2
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cupshelpers==1.0
dbus-python==1.2.12
defer==1.0.6
distro==1.3.0
distro-info===0.21ubuntu4
Django==3.0.4
django-admin==2.0.1
django-celery==3.3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-crontab==0.7.1
django-excel-base==1.0.3
django-excel-response2==3.0.1
django-filter==2.2.0
django-models-ext==1.1.9
django-rest-swagger==2.2.0
django-short-url==1.1.4
django-six==1.0.4
django-timezone-field==4.0
djangorestframework==3.11.0
duplicity==0.8.4
entrypoints==0.3
fasteners==0.12.0
furl==2.1.0
future==0.16.0
httplib2==0.11.3
idna==2.6
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.11.1
keyring==18.0.1
keyrings.alt==3.1.1
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.10.0
macaroonbakery==1.2.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.0
monotonic==1.5
mysqlclient==1.4.6
netifaces==0.10.4
oauthlib==2.1.0
olefile==0.46
openapi-codec==1.3.2
orderedmultidict==1.0.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==6.1.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.34.0
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==1.9.0+ubuntu1.3
python-crontab==2.4.0
python-dateutil==2.7.3
python-debian==0.1.36
pytz==2019.2
pyxdg==0.25
PyYAML==5.1.2
redis==3.4.1
reportlab==3.5.23
requests==2.21.0
requests-unixsocket==0.1.5
screen==1.0.1
SecretStorage==2.3.1
shortuuid==1.0.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
TimeConvert==1.5.1
tzlocal==2.0.0
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.24.1
usb-creator==0.3.7
vine==1.3.0
wadllib==1.3.3
xkit==0.0.0
xlwt==1.3.0
zipp==3.1.0
